Makes NodeContractor a proper class.
* adjusts some access modifiers so NodeContractor can be moved out of PrepareContractionHierarchies

* adds some documentation

* otherwise does not change the code
