+++ core/src/main/java/com/graphhopper/routing/ch/PrepareContractionHierarchies.java
@@ -20,7 +20,6 @@ package com.graphhopper.routing.ch;
 import com.graphhopper.coll.GHTreeMapComposed;
 import com.graphhopper.routing.*;
 import com.graphhopper.routing.util.*;
-import com.graphhopper.routing.weighting.AbstractWeighting;
 import com.graphhopper.routing.weighting.Weighting;
 import com.graphhopper.storage.*;
 import com.graphhopper.util.*;
@@ -220,7 +219,7 @@ public class PrepareContractionHierarchies extends AbstractAlgoPreparation imple
         sortedNodes = new GHTreeMapComposed();
         oldPriorities = new int[prepareGraph.getNodes()];
 
-        nodeContractor._initFromGraph(prepareFlagEncoder);
+        nodeContractor.initFromGraph(prepareFlagEncoder);
         return this;
     }
 
@@ -304,7 +303,7 @@ public class PrepareContractionHierarchies extends AbstractAlgoPreparation imple
                 logger.info(Helper.nf(counter) + ", updates:" + updateCounter
                         + ", nodes: " + Helper.nf(sortedNodes.getSize())
                         + ", shortcuts:" + Helper.nf(nodeContractor.getNewShortcuts())
-                        + ", dijkstras:" + Helper.nf(nodeContractor._getDijkstraCount())
+                        + ", dijkstras:" + Helper.nf(nodeContractor.getDijkstraCount())
                         + ", " + getTimesAsString()
                         + ", meanDegree:" + (long) meanDegree
                         + ", algo:" + nodeContractor.getPrepareAlgoMemoryUsage()
@@ -334,6 +333,8 @@ public class PrepareContractionHierarchies extends AbstractAlgoPreparation imple
             // contract node v!
             nodeContractor.setMaxVisitedNodes((int) meanDegree * 100);
             long tmpDegreeCounter = nodeContractor.contractNode(polledNode);
+            // todo: so far tmpDegreeCounter is not really the degree of the polled node but rather half of it, see todo
+            // in NodeContractor#findShortcuts
             // sliding mean value when using "*2" => slower changes
             meanDegree = (meanDegree * 2 + tmpDegreeCounter) / 3;
             // meanDegree = (meanDegree + tmpDegreeCounter) / 2;
@@ -380,7 +381,7 @@ public class PrepareContractionHierarchies extends AbstractAlgoPreparation imple
         logger.info("took:" + (int) allSW.stop().getSeconds()
                 + ", new shortcuts: " + Helper.nf(nodeContractor.getNewShortcuts())
                 + ", " + prepareWeighting
-                + ", dijkstras:" + nodeContractor._getDijkstraCount()
+                + ", dijkstras:" + nodeContractor.getDijkstraCount()
                 + ", " + getTimesAsString()
                 + ", meanDegree:" + (long) meanDegree
                 + ", initSize:" + initSize
@@ -391,13 +392,13 @@ public class PrepareContractionHierarchies extends AbstractAlgoPreparation imple
     }
 
     public void close() {
-        nodeContractor._close();
+        nodeContractor.close();
         sortedNodes = null;
         oldPriorities = null;
     }
 
     public long getDijkstraCount() {
-        return nodeContractor._getDijkstraCount();
+        return nodeContractor.getDijkstraCount();
     }
 
     public int getShortcuts() {
@@ -432,7 +433,7 @@ public class PrepareContractionHierarchies extends AbstractAlgoPreparation imple
     }
 
     Set<NodeContractor.Shortcut> testFindShortcuts(int node) {
-        return nodeContractor._testFindShortcuts(node);
+        return nodeContractor.testFindShortcuts(node);
     }
 
     int testAddShortcuts(List<NodeContractor.Shortcut> shortcuts) {
@@ -446,7 +447,7 @@ public class PrepareContractionHierarchies extends AbstractAlgoPreparation imple
      * the priority(v). Otherwise updating the priority before contracting in contractNodes() could
      * lead to a slowish or even endless loop.
      */
-    int calculatePriority(int v) {
+    private int calculatePriority(int v) {
         nodeContractor.setMaxVisitedNodes((int) meanDegree * 100);
         NodeContractor.CalcShortcutsResult calcShortcutsResult = nodeContractor.calcShortcutCount(v);
 
@@ -489,409 +490,4 @@ public class PrepareContractionHierarchies extends AbstractAlgoPreparation imple
         return "prepare|dijkstrabi|ch";
     }
 
-    static class NodeContractor {
-        private final CHGraph prepareGraph;
-        private final PreparationWeighting prepareWeighting;
-        private final TraversalMode traversalMode;
-        private final DataAccess originalEdges;
-        private final Map<Shortcut, Shortcut> shortcuts = new HashMap<>();
-        private final AddShortcutHandler addScHandler = new AddShortcutHandler();
-        private final CalcShortcutHandler calcScHandler = new CalcShortcutHandler();
-        private CHEdgeExplorer vehicleInExplorer;
-        private CHEdgeExplorer vehicleOutExplorer;
-        private IgnoreNodeFilter ignoreNodeFilter;
-        private DijkstraOneToMany prepareAlgo;
-        private int newShortcuts;
-        private long dijkstraCount;
-        private int maxVisitedNodes;
-        private StopWatch dijkstraSW = new StopWatch();
-        private int maxEdgesCount;
-        private int maxLevel;
-
-        NodeContractor(Directory dir, CHGraph prepareGraph, Weighting weighting, TraversalMode traversalMode) {
-            this.prepareGraph = prepareGraph;
-            this.prepareWeighting = new PreparationWeighting(weighting);
-            this.traversalMode = traversalMode;
-            originalEdges = dir.find("original_edges_" + AbstractWeighting.weightingToFileName(weighting));
-            originalEdges.create(1000);
-        }
-
-        private void _initFromGraph(FlagEncoder prepareFlagEncoder) {
-            maxLevel = prepareGraph.getNodes() + 1;
-            ignoreNodeFilter = new IgnoreNodeFilter(prepareGraph, maxLevel);
-            vehicleInExplorer = prepareGraph.createEdgeExplorer(new DefaultEdgeFilter(prepareFlagEncoder, true, false));
-            vehicleOutExplorer = prepareGraph.createEdgeExplorer(new DefaultEdgeFilter(prepareFlagEncoder, false, true));
-            prepareAlgo = new DijkstraOneToMany(prepareGraph, prepareWeighting, traversalMode);
-        }
-
-        private void _close() {
-            prepareAlgo.close();
-            originalEdges.close();
-        }
-
-        private void setMaxEdgesCount(int maxEdgesCount) {
-            this.maxEdgesCount = maxEdgesCount;
-        }
-
-        private void setMaxVisitedNodes(int maxVisitedNodes) {
-            this.maxVisitedNodes = maxVisitedNodes;
-        }
-
-        private long contractNode(int node) {
-            shortcuts.clear();
-            long tmpDegreeCounter = findShortcuts(addScHandler.setNode(node));
-            newShortcuts += addShortcuts(shortcuts.keySet());
-            return tmpDegreeCounter;
-        }
-
-        private CalcShortcutsResult calcShortcutCount(int v) {
-            findShortcuts(calcScHandler.setNode(v));
-            return calcScHandler.calcShortcutsResult;
-        }
-
-        /**
-         * Finds shortcuts, does not change the underlying graph.
-         */
-        long findShortcuts(ShortcutHandler sch) {
-            long tmpDegreeCounter = 0;
-            EdgeIterator incomingEdges = vehicleInExplorer.setBaseNode(sch.getNode());
-            // collect outgoing nodes (goal-nodes) only once
-            while (incomingEdges.next()) {
-                int u_fromNode = incomingEdges.getAdjNode();
-                // accept only uncontracted nodes
-                if (prepareGraph.getLevel(u_fromNode) != maxLevel)
-                    continue;
-
-                double v_u_dist = incomingEdges.getDistance();
-                double v_u_weight = prepareWeighting.calcWeight(incomingEdges, true, EdgeIterator.NO_EDGE);
-                int incomingEdge = incomingEdges.getEdge();
-                int incomingEdgeOrigCount = getOrigEdgeCount(incomingEdge);
-                // collect outgoing nodes (goal-nodes) only once
-                EdgeIterator outgoingEdges = vehicleOutExplorer.setBaseNode(sch.getNode());
-                // force fresh maps etc as this cannot be determined by from node alone (e.g. same from node but different avoidNode)
-                prepareAlgo.clear();
-                tmpDegreeCounter++;
-                while (outgoingEdges.next()) {
-                    int w_toNode = outgoingEdges.getAdjNode();
-                    // add only uncontracted nodes
-                    if (prepareGraph.getLevel(w_toNode) != maxLevel || u_fromNode == w_toNode)
-                        continue;
-
-                    // Limit weight as ferries or forbidden edges can increase local search too much.
-                    // If we decrease the correct weight we only explore less and introduce more shortcuts.
-                    // I.e. no change to accuracy is made.
-                    double existingDirectWeight = v_u_weight + prepareWeighting.calcWeight(outgoingEdges, false, incomingEdges.getEdge());
-                    if (Double.isNaN(existingDirectWeight))
-                        throw new IllegalStateException("Weighting should never return NaN values"
-                                + ", in:" + getCoords(incomingEdges, prepareGraph) + ", out:" + getCoords(outgoingEdges, prepareGraph)
-                                + ", dist:" + outgoingEdges.getDistance());
-
-                    if (Double.isInfinite(existingDirectWeight))
-                        continue;
-
-                    double existingDistSum = v_u_dist + outgoingEdges.getDistance();
-                    prepareAlgo.setWeightLimit(existingDirectWeight);
-                    prepareAlgo.setMaxVisitedNodes(maxVisitedNodes);
-                    prepareAlgo.setEdgeFilter(ignoreNodeFilter.setAvoidNode(sch.getNode()));
-
-                    dijkstraSW.start();
-                    dijkstraCount++;
-                    int endNode = prepareAlgo.findEndNode(u_fromNode, w_toNode);
-                    dijkstraSW.stop();
-
-                    // compare end node as the limit could force dijkstra to finish earlier
-                    if (endNode == w_toNode && prepareAlgo.getWeight(endNode) <= existingDirectWeight)
-                        // FOUND witness path, so do not add shortcut
-                        continue;
-
-                    sch.foundShortcut(u_fromNode, w_toNode,
-                            existingDirectWeight, existingDistSum,
-                            outgoingEdges.getEdge(), getOrigEdgeCount(outgoingEdges.getEdge()),
-                            incomingEdge, incomingEdgeOrigCount);
-                }
-            }
-            return tmpDegreeCounter;
-        }
-
-        /**
-         * Introduces the necessary shortcuts for adjNode v in the graph.
-         */
-        int addShortcuts(Collection<Shortcut> shortcuts) {
-            int tmpNewShortcuts = 0;
-            NEXT_SC:
-            for (Shortcut sc : shortcuts) {
-                boolean updatedInGraph = false;
-                // check if we need to update some existing shortcut in the graph
-                CHEdgeIterator iter = vehicleOutExplorer.setBaseNode(sc.from);
-                while (iter.next()) {
-                    if (iter.isShortcut() && iter.getAdjNode() == sc.to) {
-                        int status = iter.getMergeStatus(sc.flags);
-                        if (status == 0)
-                            continue;
-
-                        if (sc.weight >= prepareWeighting.calcWeight(iter, false, EdgeIterator.NO_EDGE)) {
-                            // special case if a bidirectional shortcut has worse weight and still has to be added as otherwise the opposite direction would be missing
-                            // see testShortcutMergeBug
-                            if (status == 2)
-                                break;
-
-                            continue NEXT_SC;
-                        }
-
-                        if (iter.getEdge() == sc.skippedEdge1 || iter.getEdge() == sc.skippedEdge2) {
-                            throw new IllegalStateException("Shortcut cannot update itself! " + iter.getEdge()
-                                    + ", skipEdge1:" + sc.skippedEdge1 + ", skipEdge2:" + sc.skippedEdge2
-                                    + ", edge " + iter + ":" + getCoords(iter, prepareGraph)
-                                    + ", sc:" + sc
-                                    + ", skippedEdge1: " + getCoords(prepareGraph.getEdgeIteratorState(sc.skippedEdge1, sc.from), prepareGraph)
-                                    + ", skippedEdge2: " + getCoords(prepareGraph.getEdgeIteratorState(sc.skippedEdge2, sc.to), prepareGraph)
-                                    + ", neighbors:" + GHUtility.getNeighbors(iter));
-                        }
-
-                        // note: flags overwrite weight => call first
-                        iter.setFlags(sc.flags);
-                        iter.setWeight(sc.weight);
-                        iter.setDistance(sc.dist);
-                        iter.setSkippedEdges(sc.skippedEdge1, sc.skippedEdge2);
-                        setOrigEdgeCount(iter.getEdge(), sc.originalEdges);
-                        updatedInGraph = true;
-                        break;
-                    }
-                }
-
-                if (!updatedInGraph) {
-                    CHEdgeIteratorState edgeState = prepareGraph.shortcut(sc.from, sc.to);
-                    // note: flags overwrite weight => call first
-                    edgeState.setFlags(sc.flags);
-                    edgeState.setWeight(sc.weight);
-                    edgeState.setDistance(sc.dist);
-                    edgeState.setSkippedEdges(sc.skippedEdge1, sc.skippedEdge2);
-                    setOrigEdgeCount(edgeState.getEdge(), sc.originalEdges);
-                    tmpNewShortcuts++;
-                }
-            }
-            return tmpNewShortcuts;
-        }
-
-        String getCoords(EdgeIteratorState e, Graph g) {
-            NodeAccess na = g.getNodeAccess();
-            int base = e.getBaseNode();
-            int adj = e.getAdjNode();
-            return base + "->" + adj + " (" + e.getEdge() + "); "
-                    + na.getLat(base) + "," + na.getLon(base) + " -> " + na.getLat(adj) + "," + na.getLon(adj);
-        }
-
-        private int getNewShortcuts() {
-            return newShortcuts;
-        }
-
-        private void setOrigEdgeCount(int edgeId, int value) {
-            edgeId -= maxEdgesCount;
-            if (edgeId < 0) {
-                // ignore setting as every normal edge has original edge count of 1
-                if (value != 1)
-                    throw new IllegalStateException("Trying to set original edge count for normal edge to a value = " + value
-                            + ", edge:" + (edgeId + maxEdgesCount) + ", max:" + maxEdgesCount + ", graph.max:");
-                // todo: is it ok to remove the following line from the error message ?, ghStorage is not available in node contractor atm
-                //+ ghStorage.getAllEdges().getMaxId());
-                return;
-            }
-
-            long tmp = (long) edgeId * 4;
-            originalEdges.ensureCapacity(tmp + 4);
-            originalEdges.setInt(tmp, value);
-        }
-
-        private int getOrigEdgeCount(int edgeId) {
-            edgeId -= maxEdgesCount;
-            if (edgeId < 0)
-                return 1;
-
-            long tmp = (long) edgeId * 4;
-            originalEdges.ensureCapacity(tmp + 4);
-            return originalEdges.getInt(tmp);
-        }
-
-        private Set<Shortcut> _testFindShortcuts(int node) {
-            findShortcuts(addScHandler.setNode(node));
-            return shortcuts.keySet();
-        }
-
-        private String getPrepareAlgoMemoryUsage() {
-            return prepareAlgo.getMemoryUsageAsString();
-        }
-
-        private long _getDijkstraCount() {
-            return dijkstraCount;
-        }
-
-        private void resetDijkstraTime() {
-            dijkstraSW = new StopWatch();
-        }
-
-        private float getDijkstraSeconds() {
-            return dijkstraSW.getSeconds();
-        }
-
-        static class IgnoreNodeFilter implements EdgeFilter {
-            int avoidNode;
-            int maxLevel;
-            CHGraph graph;
-
-            public IgnoreNodeFilter(CHGraph g, int maxLevel) {
-                this.graph = g;
-                this.maxLevel = maxLevel;
-            }
-
-            public IgnoreNodeFilter setAvoidNode(int node) {
-                this.avoidNode = node;
-                return this;
-            }
-
-            @Override
-            public final boolean accept(EdgeIteratorState iter) {
-                // ignore if it is skipNode or adjNode is already contracted
-                int node = iter.getAdjNode();
-                return avoidNode != node && graph.getLevel(node) == maxLevel;
-            }
-        }
-
-        static class Shortcut {
-            int from;
-            int to;
-            int skippedEdge1;
-            int skippedEdge2;
-            double dist;
-            double weight;
-            int originalEdges;
-            long flags = PrepareEncoder.getScFwdDir();
-
-            public Shortcut(int from, int to, double weight, double dist) {
-                this.from = from;
-                this.to = to;
-                this.weight = weight;
-                this.dist = dist;
-            }
-
-            @Override
-            public int hashCode() {
-                int hash = 5;
-                hash = 23 * hash + from;
-                hash = 23 * hash + to;
-                return 23 * hash
-                        + (int) (Double.doubleToLongBits(this.weight) ^ (Double.doubleToLongBits(this.weight) >>> 32));
-            }
-
-            @Override
-            public boolean equals(Object obj) {
-                if (obj == null || getClass() != obj.getClass())
-                    return false;
-
-                final Shortcut other = (Shortcut) obj;
-                if (this.from != other.from || this.to != other.to)
-                    return false;
-
-                return Double.doubleToLongBits(this.weight) == Double.doubleToLongBits(other.weight);
-            }
-
-            @Override
-            public String toString() {
-                String str;
-                if (flags == PrepareEncoder.getScDirMask())
-                    str = from + "<->";
-                else
-                    str = from + "->";
-
-                return str + to + ", weight:" + weight + " (" + skippedEdge1 + "," + skippedEdge2 + ")";
-            }
-        }
-
-        interface ShortcutHandler {
-            void foundShortcut(int u_fromNode, int w_toNode,
-                               double existingDirectWeight, double distance,
-                               int outgoingEdge, int outgoingEdgeOrigCount,
-                               int incomingEdge, int incomingEdgeOrigCount);
-
-            int getNode();
-        }
-
-        class CalcShortcutHandler implements ShortcutHandler {
-            int node;
-            CalcShortcutsResult calcShortcutsResult = new CalcShortcutsResult();
-
-            @Override
-            public int getNode() {
-                return node;
-            }
-
-            public CalcShortcutHandler setNode(int n) {
-                node = n;
-                calcShortcutsResult.originalEdgesCount = 0;
-                calcShortcutsResult.shortcutsCount = 0;
-                return this;
-            }
-
-            @Override
-            public void foundShortcut(int u_fromNode, int w_toNode,
-                                      double existingDirectWeight, double distance,
-                                      int outgoingEdge, int outgoingEdgeOrigCount,
-                                      int incomingEdge, int incomingEdgeOrigCount) {
-                calcShortcutsResult.shortcutsCount++;
-                calcShortcutsResult.originalEdgesCount += incomingEdgeOrigCount + outgoingEdgeOrigCount;
-            }
-        }
-
-        class AddShortcutHandler implements ShortcutHandler {
-            int node;
-
-            @Override
-            public int getNode() {
-                return node;
-            }
-
-            public AddShortcutHandler setNode(int n) {
-                shortcuts.clear();
-                node = n;
-                return this;
-            }
-
-            @Override
-            public void foundShortcut(int u_fromNode, int w_toNode,
-                                      double existingDirectWeight, double existingDistSum,
-                                      int outgoingEdge, int outgoingEdgeOrigCount,
-                                      int incomingEdge, int incomingEdgeOrigCount) {
-                // FOUND shortcut
-                // but be sure that it is the only shortcut in the collection
-                // and also in the graph for u->w. If existing AND identical weight => update setProperties.
-                // Hint: shortcuts are always one-way due to distinct level of every node but we don't
-                // know yet the levels so we need to determine the correct direction or if both directions
-                Shortcut sc = new Shortcut(u_fromNode, w_toNode, existingDirectWeight, existingDistSum);
-                if (shortcuts.containsKey(sc))
-                    return;
-
-                Shortcut tmpSc = new Shortcut(w_toNode, u_fromNode, existingDirectWeight, existingDistSum);
-                Shortcut tmpRetSc = shortcuts.get(tmpSc);
-                if (tmpRetSc != null) {
-                    // overwrite flags only if skipped edges are identical
-                    if (tmpRetSc.skippedEdge2 == incomingEdge && tmpRetSc.skippedEdge1 == outgoingEdge) {
-                        tmpRetSc.flags = PrepareEncoder.getScDirMask();
-                        return;
-                    }
-                }
-
-                Shortcut old = shortcuts.put(sc, sc);
-                if (old != null)
-                    throw new IllegalStateException("Shortcut did not exist (" + sc + ") but was overwriting another one? " + old);
-
-                sc.skippedEdge1 = incomingEdge;
-                sc.skippedEdge2 = outgoingEdge;
-                sc.originalEdges = incomingEdgeOrigCount + outgoingEdgeOrigCount;
-            }
-        }
-
-        static class CalcShortcutsResult {
-            int originalEdgesCount;
-            int shortcutsCount;
-        }
-    }
 }
